fix: correctly sort books with a half-star rating.
When sorting by rating, books with a rating of a half-star aren't sorted
correctly. This was due to the ratings being converted to strings and
those strings used for sorting. A half-star is stored as 5, which means
that the string gets sorted as being above anything starting with 4 or
less. I.e. anything but a full five stars. This has been fixed by doing
the comparison on the int directly.